Dont try to delete our temp dir if it doesn't exist yet.

+++ installcab.py
@@ -10,7 +10,8 @@ import shutil
 # Utils
 
 def cleanup():
-    shutil.rmtree(tmpdir)
+    if os.path.exists(tmpdir):
+        shutil.rmtree(tmpdir)
 
 def bad_exit(text):
     print(text)
